Add Pull Request Template (#1)
# Pull Request

## Description

This pull request introduces a new template for pull requests to ensure
consistency and completeness in the descriptions. The most important
change is the addition of a structured format to the
`.github/pull_request_template.md` file.

Documentation improvements:

*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bR1-R7):
Added a new pull request template with sections for description, issue
reference, motivation, context, and dependencies.