Add Code of Conduct (#4)
# Pull Request

## Description

This pull request adds a new `CODE_OF_CONDUCT.md` file to the `docs`
directory. The document outlines the expected behavior and
responsibilities of community members, as well as the enforcement
mechanisms for addressing violations.

Key additions include:

* Introduction of the Contributor Covenant Code of Conduct to ensure a
harassment-free experience for all community members.
* Detailed standards of acceptable and unacceptable behavior within the
community.
* Responsibilities of community leaders in enforcing the code of conduct
and handling violations.
* Guidelines for enforcement actions, ranging from warnings to permanent
bans, based on the severity of the violation.
* Attribution and references to the original Contributor Covenant and
related resources.