Add Dependency Review (#12)
# Pull Request

## Description

This pull request includes a few significant changes to the GitHub
Actions workflow configuration. The most important change is the
addition of a new job for dependency review.

Enhancements to GitHub Actions workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR37-R48):
Added a new job named `dependency-review` which includes steps for
checking out the repository and running the dependency review action.