Add Zizmor Checking (#14)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow in
the `.github/workflows/code-checks.yml` file. The job is designed to
check GitHub Actions using the `zizmor` tool.

Key changes include:

* **New job addition:**
*
[`run-zizmor`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R52-R76):
Added a new job named `run-zizmor` to the workflow, which runs on
`ubuntu-latest` and includes several steps to checkout the repository,
install the latest version of `uv`, run `zizmor`, and upload the SARIF
file.