Fix Security Policy Links (#26)
# Pull Request

## Description

This pull request includes updates to the `docs/SECURITY.md` file to
correct repository links. The changes ensure that the links point to the
correct repository for reporting vulnerabilities and viewing security
advisories.

Updates to security documentation:

*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9):
Updated the link in the "To report a vulnerability" section to point to
the correct repository's Security tab.
*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L23-R23):
Updated the link in the "Security Updates" section to point to the
correct repository's Security Advisories section.