Add CodeQL configuration (#27)
# Pull Request

## Description

This pull request includes changes to the CodeQL configuration and
workflow to improve the security and quality of the code analysis. The
most important changes involve excluding specific queries and updating
the workflow to use the new configuration file.

Updates to CodeQL configuration and workflow:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beR1-R3):
Added a query filter to exclude the `actions/unpinned-tag` query.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R96):
Updated the CodeQL analysis job to use the new configuration file
`.github/other-configurations/codeql-config.yml`.