Check Justfile Format (#31)
# Pull Request

## Description

This pull request introduces a new GitHub Actions job for checking the
format of the Justfile and adds several commands to the `Justfile` for
formatting and other tasks.

### GitHub Actions:

* Added a new job `check-justfile-format` to
`.github/workflows/code-checks.yml` to ensure the Justfile is properly
formatted. This job uses the `extractions/setup-just` action and runs
`just format-check`.

### Justfile:

* Added commands for Prettier to check and format code (`prettier-check`
and `prettier-format`).
* Added commands for Justfile formatting (`format` and `format-check`).
* Added a command for detecting secrets using `gitleaks`
(`gitleaks-detect`).
* Added commands to install Git hooks for pre-commit and post-commit
actions (`install-git-hooks`).