Update Labels (#36)
# Pull Request

## Description

This pull request includes updates to configuration files to improve
consistency and clarity. The most important changes are the renaming of
configuration-related labels and sections.

Configuration updates:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L12-R12):
Renamed the `configuration` section to `configurations` for consistency.
*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dL6-R7):
Updated the `configuration` label to `configurations` and revised its
description to be more specific about the changes it represents.