Add GitHub Action Step Name (#57)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflow and
the Lefthook configuration file. The changes improve label
synchronization and ensure compatibility with the latest Lefthook
features.

### Workflow improvements:
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L25-R26):
Added a descriptive name ("Sync labels") to the label synchronization
step for better clarity in the GitHub Actions workflow.

### Lefthook configuration update:
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` requirement from `1.11.4` to `1.11.11` to
ensure compatibility with newer Lefthook features.