Fix Zizmor SARIF Command (#69)
# Pull Request

## Description

This pull request updates the workflow configuration and introduces new
commands in the `Justfile` to streamline code checks and improve
maintainability. The key changes include upgrading the `uv` setup
action, adding a new setup step for `Just`, and refactoring the commands
for running `zizmor` checks.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L83-R87):
Upgraded the `astral-sh/setup-uv` action to version `6.0.1` and added a
new step to set up `Just` using the `extractions/setup-just` action.
Additionally, the `zizmor` command was updated to use the `just
zizmor-check-sarif` alias for generating SARIF output.

### Command refactoring:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL47-R51):
Refactored the `zizmor-check` command to use `uvx` and introduced a new
`zizmor-check-sarif` command for generating SARIF output with the
`zizmor` tool.