Bump GitHub Actions (#71)
# Pull Request

## Description

This pull request updates the versions of third-party GitHub Actions
used in the workflows to ensure compatibility and access to the latest
features and fixes.

Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L106-R106):
Updated the `setup-uv` action from version `v5.4.0` to `v6.0.1` to
incorporate the latest improvements and fixes.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL50-R50):
Updated the `dependency-review-action` from version `v4.6.0` to `v4.7.0`
to ensure the latest dependency review features are utilized.