Remove `.editorconfig-checker.json` (#81)
# Pull Request

## Description

This pull request removes the configuration file for the
`editorconfig-checker` tool and updates the GitHub Actions workflow to
reflect this change.

### Configuration Removal:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Deleted the `editorconfig-checker` configuration file, which included
settings such as verbosity, debugging, and file exclusions.

### Workflow Update:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L36):
Removed the `EDITORCONFIG_FILE_NAME` environment variable, as the
`editorconfig-checker` configuration file is no longer required.